fix(sdk): isolate startup diagnostics from broker internals
Startup marker helpers must not pull the broker implementation into SDK client entrypoints. Keep the process completion helper isolated and move shared diagnostic serialization to its own dependency-safe module.

Lore-id: 4015-sdk-closure\nConfidence: high\nScope-risk: narrow\nReversibility: straightforward\nTested: coding-agent check; sdk-broker-lifecycle-cleanup
